app/requirements.txt
nicegui==2.23.3
GitPython==3.1.45